attach_issue

This function is used to attach an issue to either

an external object (in which case you must specify the external_object_doi argument)

a data product (in which case you must specify the namespace and data_product arguments), or

a component within a data product (in which case you must specify the namespace, data_product, and component arguments)

attach_issue(
  description,
  severity,
  external_object_doi,
  namespace,
  data_product,
  component,
  version,
  key
)

Arguments

description
a string containing a free text description of the issue

severity
an integer specifying the severity of the issue

external_object_doi
a string specifying the DOI or name of the external_object

namespace
a string specifying the name of the namespace

data_product
a string specifying the name of the data_product

component
a string specifying the name of the object_component, unique in the context of object_component and its object reference

version
a string specifying the version number of the data_product or external_object

key
API token from data.scrc.uk

Examples