create_distribution

Function to populate toml file with distribution type data.

create_distribution(filename, path, name, distribution, parameters)

Arguments

filename
a string specifying the name of the toml file, e.g. "0.1.0.toml"

path
a string specifying the directory in which you want to save the toml file

name
a string specifying the name of the parameter

distribution
a string specifying the name of the distribution

parameters
a named list specifying parameter values and associated names (see example)

Examples

filename <- "test.toml"
path <- "data-raw/example"
name <- "latency"
distribution = "gamma"
parameters <- list(shape = 2.0, scale = 3.0)
create_distribution(filename, path, name, distribution, parameters)