get_entry

get_entry

get_entry(table, query)

Arguments

table
a string specifying the name of the table

query
a list containing a valid query for the table, e.g. list(field = value)

Value

Returns a list of fields present in the specified entry

Examples

# Get list of entries
data_product <- "records/SARS-CoV-2/scotland/cases-and-management/ambulance"
get_entry("data_product", list(name = data_product))

#> [[1]]
#> [[1]]$url
#> [1] "https://data.scrc.uk/api/data_product/2285/"
#> 
#> [[1]]$last_updated
#> [1] "2020-10-28T14:30:23.285702Z"
#> 
#> [[1]]$name
#> [1] "records/SARS-CoV-2/scotland/cases-and-management/ambulance"
#> 
#> [[1]]$version
#> [1] "0.1.6"
#> 
#> [[1]]$updated_by
#> [1] "https://data.scrc.uk/api/users/3/"
#> 
#> [[1]]$object
#> [1] "https://data.scrc.uk/api/object/68407/"
#> 
#> [[1]]$namespace
#> [1] "https://data.scrc.uk/api/namespace/2/"
#> 
#> 
#> [[2]]
#> [[2]]$url
#> [1] "https://data.scrc.uk/api/data_product/2233/"
#> 
#> [[2]]$last_updated
#> [1] "2020-10-27T14:30:09.305228Z"
#> 
#> [[2]]$name
#> [1] "records/SARS-CoV-2/scotland/cases-and-management/ambulance"
#> 
#> [[2]]$version
#> [1] "0.1.5"
#> 
#> [[2]]$updated_by
#> [1] "https://data.scrc.uk/api/users/3/"
#> 
#> [[2]]$object
#> [1] "https://data.scrc.uk/api/object/67982/"
#> 
#> [[2]]$namespace
#> [1] "https://data.scrc.uk/api/namespace/2/"
#> 
#> 
#> [[3]]
#> [[3]]$url
#> [1] "https://data.scrc.uk/api/data_product/2225/"
#> 
#> [[3]]$last_updated
#> [1] "2020-10-26T14:30:30.461496Z"
#> 
#> [[3]]$name
#> [1] "records/SARS-CoV-2/scotland/cases-and-management/ambulance"
#> 
#> [[3]]$version
#> [1] "0.1.4"
#> 
#> [[3]]$updated_by
#> [1] "https://data.scrc.uk/api/users/3/"
#> 
#> [[3]]$object
#> [1] "https://data.scrc.uk/api/object/67972/"
#> 
#> [[3]]$namespace
#> [1] "https://data.scrc.uk/api/namespace/2/"
#> 
#> 
#> [[4]]
#> [[4]]$url
#> [1] "https://data.scrc.uk/api/data_product/2199/"
#> 
#> [[4]]$last_updated
#> [1] "2020-10-25T14:30:08.639983Z"
#> 
#> [[4]]$name
#> [1] "records/SARS-CoV-2/scotland/cases-and-management/ambulance"
#> 
#> [[4]]$version
#> [1] "0.1.3"
#> 
#> [[4]]$updated_by
#> [1] "https://data.scrc.uk/api/users/3/"
#> 
#> [[4]]$object
#> [1] "https://data.scrc.uk/api/object/67800/"
#> 
#> [[4]]$namespace
#> [1] "https://data.scrc.uk/api/namespace/2/"
#> 
#> 
#> [[5]]
#> [[5]]$url
#> [1] "https://data.scrc.uk/api/data_product/2191/"
#> 
#> [[5]]$last_updated
#> [1] "2020-10-24T13:30:20.829398Z"
#> 
#> [[5]]$name
#> [1] "records/SARS-CoV-2/scotland/cases-and-management/ambulance"
#> 
#> [[5]]$version
#> [1] "0.1.2"
#> 
#> [[5]]$updated_by
#> [1] "https://data.scrc.uk/api/users/3/"
#> 
#> [[5]]$object
#> [1] "https://data.scrc.uk/api/object/67790/"
#> 
#> [[5]]$namespace
#> [1] "https://data.scrc.uk/api/namespace/2/"
#> 
#> 
#> [[6]]
#> [[6]]$url
#> [1] "https://data.scrc.uk/api/data_product/2096/"
#> 
#> [[6]]$last_updated
#> [1] "2020-10-23T16:13:03.052122Z"
#> 
#> [[6]]$name
#> [1] "records/SARS-CoV-2/scotland/cases-and-management/ambulance"
#> 
#> [[6]]$version
#> [1] "0.1.1"
#> 
#> [[6]]$updated_by
#> [1] "https://data.scrc.uk/api/users/3/"
#> 
#> [[6]]$object
#> [1] "https://data.scrc.uk/api/object/66997/"
#> 
#> [[6]]$namespace
#> [1] "https://data.scrc.uk/api/namespace/2/"
#> 
#> 
#> [[7]]
#> [[7]]$url
#> [1] "https://data.scrc.uk/api/data_product/347/"
#> 
#> [[7]]$last_updated
#> [1] "2020-09-08T16:18:46.423762Z"
#> 
#> [[7]]$name
#> [1] "records/SARS-CoV-2/scotland/cases-and-management/ambulance"
#> 
#> [[7]]$version
#> [1] "0.1.0"
#> 
#> [[7]]$updated_by
#> [1] "https://data.scrc.uk/api/users/3/"
#> 
#> [[7]]$object
#> [1] "https://data.scrc.uk/api/object/45430/"
#> 
#> [[7]]$namespace
#> [1] "https://data.scrc.uk/api/namespace/2/"
#> 
#>