Get existing entries

Get entries (from the data registry) in a particular table

get_existing(table, limit_results = TRUE, detail = "all")

Arguments

table
a string specifying the name of the table

limit_results
a boolean specifying whether or not to limit the results, default is TRUE

detail
a string specifying what level of detail to return; options are "all" for all details or "id" for just URL and IDs

Value

Returns a data.frame of entries in table, default is limited to 100 entries

See also

Examples

# Get all entries in storage_root
get_existing("storage_root", limit_results = FALSE)

#> # A tibble: 78 x 7
#>    url     last_updated  name     root    accessibility updated_by  locations   
#>    <chr>   <chr>         <chr>    <chr>   <chr>         <chr>       <chr>       
#>  1 https:… 2020-10-30T1… source-… source… 0             https://da… https://dat…
#>  2 https:… 2020-10-30T1… https:/… https:… 0             https://da… https://dat…
#>  3 https:… 2020-10-30T1… source … https:… 0             https://da… https://dat…
#>  4 https:… 2020-10-30T1… source-… source… 0             https://da… https://dat…
#>  5 https:… 2020-10-30T1… https:/… https:… 0             https://da… https://dat…
#>  6 https:… 2020-10-30T1… source … https:… 0             https://da… https://dat…
#>  7 https:… 2020-10-30T1… source-… source… 0             https://da… https://dat…
#>  8 https:… 2020-10-30T1… https:/… https:… 0             https://da… https://dat…
#>  9 https:… 2020-10-30T1… source-… source… 0             https://da… https://dat…
#> 10 https:… 2020-10-30T1… https:/… https:… 0             https://da… https://dat…
#> # … with 68 more rows

# Get 100 newest entries in data_product
get_existing("data_product")

#> # A tibble: 100 x 7
#>    url       last_updated   name      version updated_by   object    namespace  
#>    <chr>     <chr>          <chr>     <chr>   <chr>        <chr>     <chr>      
#>  1 https://… 2020-10-30T16… register… 7.2020… https://dat… https://… https://da…
#>  2 https://… 2020-10-30T16… data_pro… 0.1.0   https://dat… https://… https://da…
#>  3 https://… 2020-10-30T16… register… 7.2020… https://dat… https://… https://da…
#>  4 https://… 2020-10-30T16… data_pro… 0.1.0   https://dat… https://… https://da…
#>  5 https://… 2020-10-30T16… register… 7.2020… https://dat… https://… https://da…
#>  6 https://… 2020-10-30T16… register… 7.2020… https://dat… https://… https://da…
#>  7 https://… 2020-10-30T16… data_pro… 0.1.0   https://dat… https://… https://da…
#>  8 https://… 2020-10-30T16… register… 1.2020… https://dat… https://… https://da…
#>  9 https://… 2020-10-30T16… data_pro… 0.1.0   https://dat… https://… https://da…
#> 10 https://… 2020-10-30T16… data_pro… 0.1.0   https://dat… https://… https://da…
#> # … with 90 more rows