Get GitHub package info

Function to retrieve the current version of a GitHub repository and the GitHub hash associated with it. The repository should contain a DESCRIPTION file whether it is set up as a package or not. If the repository is set up as a package, checks are made to ensure remote and local versions of the package have the same version number.

get_package_info(repo, script_path, package)

Arguments

repo
a string specifying the github username/repository

script_path
a string specifying the submission script filename and extension

package
(optional) a string specifying the name of the package containing the submission script

Value

Returns a list comprising:

the github storage root name: "github"

the github repository: e.g. "ScottishCovidResponse/SCRCdata"

the github repository version number:

the github repo hash

the submission script file name

Examples

# \donttest{
if (FALSE) {
get_package_info(repo = "ScottishCovidResponse/SCRCdata",
script_path = "scotgov_deaths.R",
package = "SCRCdata")
}# }