Get URL

Retrieve URLs for particular entries or all entries in a table

get_url(table, query = list())

Arguments

table
a string specifying the name of the table

query
a list containing a valid query for the table, e.g. list(field = value)

Value

Returns a string or list of strings specifying the URL or URLs of entries in a table

Examples

# Retrieve URLs for all entries in a table
get_url(table = "storage_root")

#> [[1]]
#> [1] "https://data.scrc.uk/api/storage_root/4202/"
#> 
#> [[2]]
#> [1] "https://data.scrc.uk/api/storage_root/4201/"
#> 
#> [[3]]
#> [1] "https://data.scrc.uk/api/storage_root/4200/"
#> 
#> [[4]]
#> [1] "https://data.scrc.uk/api/storage_root/4199/"
#> 
#> [[5]]
#> [1] "https://data.scrc.uk/api/storage_root/4198/"
#> 
#> [[6]]
#> [1] "https://data.scrc.uk/api/storage_root/4197/"
#> 
#> [[7]]
#> [1] "https://data.scrc.uk/api/storage_root/4196/"
#> 
#> [[8]]
#> [1] "https://data.scrc.uk/api/storage_root/4195/"
#> 
#> [[9]]
#> [1] "https://data.scrc.uk/api/storage_root/4194/"
#> 
#> [[10]]
#> [1] "https://data.scrc.uk/api/storage_root/4193/"
#> 
#> [[11]]
#> [1] "https://data.scrc.uk/api/storage_root/4192/"
#> 
#> [[12]]
#> [1] "https://data.scrc.uk/api/storage_root/4191/"
#> 
#> [[13]]
#> [1] "https://data.scrc.uk/api/storage_root/4190/"
#> 
#> [[14]]
#> [1] "https://data.scrc.uk/api/storage_root/4189/"
#> 
#> [[15]]
#> [1] "https://data.scrc.uk/api/storage_root/4188/"
#> 
#> [[16]]
#> [1] "https://data.scrc.uk/api/storage_root/4187/"
#> 
#> [[17]]
#> [1] "https://data.scrc.uk/api/storage_root/4186/"
#> 
#> [[18]]
#> [1] "https://data.scrc.uk/api/storage_root/4185/"
#> 
#> [[19]]
#> [1] "https://data.scrc.uk/api/storage_root/4184/"
#> 
#> [[20]]
#> [1] "https://data.scrc.uk/api/storage_root/4183/"
#> 
#> [[21]]
#> [1] "https://data.scrc.uk/api/storage_root/4182/"
#> 
#> [[22]]
#> [1] "https://data.scrc.uk/api/storage_root/4181/"
#> 
#> [[23]]
#> [1] "https://data.scrc.uk/api/storage_root/4180/"
#> 
#> [[24]]
#> [1] "https://data.scrc.uk/api/storage_root/4179/"
#> 
#> [[25]]
#> [1] "https://data.scrc.uk/api/storage_root/4178/"
#> 
#> [[26]]
#> [1] "https://data.scrc.uk/api/storage_root/4177/"
#> 
#> [[27]]
#> [1] "https://data.scrc.uk/api/storage_root/4176/"
#> 
#> [[28]]
#> [1] "https://data.scrc.uk/api/storage_root/4175/"
#> 
#> [[29]]
#> [1] "https://data.scrc.uk/api/storage_root/4174/"
#> 
#> [[30]]
#> [1] "https://data.scrc.uk/api/storage_root/4173/"
#> 
#> [[31]]
#> [1] "https://data.scrc.uk/api/storage_root/4172/"
#> 
#> [[32]]
#> [1] "https://data.scrc.uk/api/storage_root/4171/"
#> 
#> [[33]]
#> [1] "https://data.scrc.uk/api/storage_root/4170/"
#> 
#> [[34]]
#> [1] "https://data.scrc.uk/api/storage_root/4169/"
#> 
#> [[35]]
#> [1] "https://data.scrc.uk/api/storage_root/4168/"
#> 
#> [[36]]
#> [1] "https://data.scrc.uk/api/storage_root/4167/"
#> 
#> [[37]]
#> [1] "https://data.scrc.uk/api/storage_root/4166/"
#> 
#> [[38]]
#> [1] "https://data.scrc.uk/api/storage_root/4165/"
#> 
#> [[39]]
#> [1] "https://data.scrc.uk/api/storage_root/4164/"
#> 
#> [[40]]
#> [1] "https://data.scrc.uk/api/storage_root/4163/"
#> 
#> [[41]]
#> [1] "https://data.scrc.uk/api/storage_root/4162/"
#> 
#> [[42]]
#> [1] "https://data.scrc.uk/api/storage_root/4161/"
#> 
#> [[43]]
#> [1] "https://data.scrc.uk/api/storage_root/4160/"
#> 
#> [[44]]
#> [1] "https://data.scrc.uk/api/storage_root/4159/"
#> 
#> [[45]]
#> [1] "https://data.scrc.uk/api/storage_root/4158/"
#> 
#> [[46]]
#> [1] "https://data.scrc.uk/api/storage_root/4157/"
#> 
#> [[47]]
#> [1] "https://data.scrc.uk/api/storage_root/4156/"
#> 
#> [[48]]
#> [1] "https://data.scrc.uk/api/storage_root/4155/"
#> 
#> [[49]]
#> [1] "https://data.scrc.uk/api/storage_root/4154/"
#> 
#> [[50]]
#> [1] "https://data.scrc.uk/api/storage_root/4153/"
#> 
#> [[51]]
#> [1] "https://data.scrc.uk/api/storage_root/4152/"
#> 
#> [[52]]
#> [1] "https://data.scrc.uk/api/storage_root/4151/"
#> 
#> [[53]]
#> [1] "https://data.scrc.uk/api/storage_root/4150/"
#> 
#> [[54]]
#> [1] "https://data.scrc.uk/api/storage_root/4149/"
#> 
#> [[55]]
#> [1] "https://data.scrc.uk/api/storage_root/4148/"
#> 
#> [[56]]
#> [1] "https://data.scrc.uk/api/storage_root/4052/"
#> 
#> [[57]]
#> [1] "https://data.scrc.uk/api/storage_root/2968/"
#> 
#> [[58]]
#> [1] "https://data.scrc.uk/api/storage_root/2967/"
#> 
#> [[59]]
#> [1] "https://data.scrc.uk/api/storage_root/2966/"
#> 
#> [[60]]
#> [1] "https://data.scrc.uk/api/storage_root/870/"
#> 
#> [[61]]
#> [1] "https://data.scrc.uk/api/storage_root/465/"
#> 
#> [[62]]
#> [1] "https://data.scrc.uk/api/storage_root/463/"
#> 
#> [[63]]
#> [1] "https://data.scrc.uk/api/storage_root/410/"
#> 
#> [[64]]
#> [1] "https://data.scrc.uk/api/storage_root/408/"
#> 
#> [[65]]
#> [1] "https://data.scrc.uk/api/storage_root/406/"
#> 
#> [[66]]
#> [1] "https://data.scrc.uk/api/storage_root/405/"
#> 
#> [[67]]
#> [1] "https://data.scrc.uk/api/storage_root/373/"
#> 
#> [[68]]
#> [1] "https://data.scrc.uk/api/storage_root/372/"
#> 
#> [[69]]
#> [1] "https://data.scrc.uk/api/storage_root/299/"
#> 
#> [[70]]
#> [1] "https://data.scrc.uk/api/storage_root/298/"
#> 
#> [[71]]
#> [1] "https://data.scrc.uk/api/storage_root/203/"
#> 
#> [[72]]
#> [1] "https://data.scrc.uk/api/storage_root/56/"
#> 
#> [[73]]
#> [1] "https://data.scrc.uk/api/storage_root/15/"
#> 
#> [[74]]
#> [1] "https://data.scrc.uk/api/storage_root/14/"
#> 
#> [[75]]
#> [1] "https://data.scrc.uk/api/storage_root/11/"
#> 
#> [[76]]
#> [1] "https://data.scrc.uk/api/storage_root/10/"
#> 
#> [[77]]
#> [1] "https://data.scrc.uk/api/storage_root/9/"
#> 
#> [[78]]
#> [1] "https://data.scrc.uk/api/storage_root/1/"
#> 

# Retrieve the URL of a particular entry in a table
get_url(table = "storage_root", list(name = "github"))

#> [1] "https://data.scrc.uk/api/storage_root/11/"