register_everything

This function assumes that (1) the source data is being saved as a csv file, (2) the data product is being saved as an h5 file, (3) the source data and data product are being stored on the boydorr ftp server (4) the submission script is being stored in the ScottishCovidResponse/SCRCdata respository. Data products may be generated from a single or multiple original sources, see Examples below for details.

register_everything(
  product_name,
  version_number,
  save_location = "data-raw",
  doi_or_unique_name,
  namespace,
  submission_script,
  original_source_name,
  original_sourceId,
  original_root,
  original_path,
  source_filename,
  accessibility = 0,
  key
)

Arguments

product_name
a string specifying the product_name (used to identify the data product as well as being used to generate various file locations)

version_number
a string specifying the version identifier of the data_product (must conform to semantic versioning syntax)

save_location
(optional) a string, which when prepended to product_name, will specify where to download the original source to and where to save the processed data product to. Default is "data-raw"

doi_or_unique_name
a string specifying the DOI or name of the external_object (source data)

namespace
a string specifying the namespace; by default this is "SCRC"

submission_script
a string specifying the filename of the submission script located in the inst/namespace/ directory of the SCRCdata package

original_source_name
a string specifying the original source name

original_sourceId
a string specifying the API URL of the associated source table

original_root
a string specifying the root of the original source (must have a trailing slash), which when prepended to original_path produces a complete URL

original_path
a string specifying the path from the original_root, which when appended to original_root produces a complete URL

source_filename
a string specifying the source filename

accessibility
(optional) an integer value for the accessibility enum associated with original_root, where 0 is public (default) and 1 is private

key
API token from data.scrc.uk

Details

Note that when a data product is generated from multiple original sources, original_source_name, original_sourceId, original_root, and original_path are input as named lists (see Example below).

Examples