Upload paper metadata to the data registry

Upload paper metadata to the data registry

upload_paper(
  title,
  authors,
  journal,
  journal_abbreviation,
  journal_website,
  release_date,
  abstract,
  keywords,
  doi,
  primary_not_supplement = TRUE,
  version = "1.0.0",
  key
)

Arguments

title
a string specifying the title of the paper e.g. "Covid-19: A systemic disease treated with a wide-ranging approach: A case report"

authors
a string specifying the authors e.g. "Massabeti, Rosanna and Cipriani, Maria Stella and Valenti, Ivana"

journal
a string specifying the full journal name e.g. "Journal of Population Therapeutics and Clinical Pharmacology"

journal_abbreviation
a string specifying the journal abbreviation e.g. "J Popul Ther Clin Pharmacol"

journal_website
a string specifying the journal homepage e.g. "https://www.jptcp.com/index.php/jptcp"

release_date
a POSIXct of format "%Y-%m-%d %H:%M:%S" specifying the release date of the paper e.g. as.POSIXct("2020-01-01 12:00:00", format = "%Y-%m-%d %H:%M:%S")

abstract
a string specifying the abstract e.g. "At the end of December 2019, the Health Commission of the"

keywords
a string specifying keywords or keyphrases seperated by "and", e.g. "covid-19 and coronavirus disease and monoclonal antibodies

and non-invasive mechanical ventilation and treatment"

doi
a string specifying the doi e.g. "10.15586/jptcp.v27iSP1.691"

primary_not_supplement
(optional) an object of class logical where TRUE is the primary source (default) and false is a supplement

version
(optional) a string specifying the version number; by default this is 1.0.0 (assuming a published paper), set to 0.1.0 for a preprint

key
API token from data.scrc.uk

See also