Get URL

Retrieve URLs for particular entries or all entries in a table

get_url(table, query = list())

Arguments

table
a string specifying the name of the table

query
a list containing a valid query for the table, e.g. list(field = value)

Value

Returns a string or list of strings specifying the URL or URLs of entries in a table

Examples

# Retrieve URLs for all entries in a table
get_url(table = "storage_root")

#> [[1]]
#> [1] "https://data.scrc.uk/api/storage_root/4370/"
#> 
#> [[2]]
#> [1] "https://data.scrc.uk/api/storage_root/4369/"
#> 
#> [[3]]
#> [1] "https://data.scrc.uk/api/storage_root/4368/"
#> 
#> [[4]]
#> [1] "https://data.scrc.uk/api/storage_root/4367/"
#> 
#> [[5]]
#> [1] "https://data.scrc.uk/api/storage_root/4366/"
#> 
#> [[6]]
#> [1] "https://data.scrc.uk/api/storage_root/4365/"
#> 
#> [[7]]
#> [1] "https://data.scrc.uk/api/storage_root/4364/"
#> 
#> [[8]]
#> [1] "https://data.scrc.uk/api/storage_root/4363/"
#> 
#> [[9]]
#> [1] "https://data.scrc.uk/api/storage_root/4362/"
#> 
#> [[10]]
#> [1] "https://data.scrc.uk/api/storage_root/4361/"
#> 
#> [[11]]
#> [1] "https://data.scrc.uk/api/storage_root/4360/"
#> 
#> [[12]]
#> [1] "https://data.scrc.uk/api/storage_root/4359/"
#> 
#> [[13]]
#> [1] "https://data.scrc.uk/api/storage_root/4358/"
#> 
#> [[14]]
#> [1] "https://data.scrc.uk/api/storage_root/4357/"
#> 
#> [[15]]
#> [1] "https://data.scrc.uk/api/storage_root/4356/"
#> 
#> [[16]]
#> [1] "https://data.scrc.uk/api/storage_root/4355/"
#> 
#> [[17]]
#> [1] "https://data.scrc.uk/api/storage_root/4354/"
#> 
#> [[18]]
#> [1] "https://data.scrc.uk/api/storage_root/4353/"
#> 
#> [[19]]
#> [1] "https://data.scrc.uk/api/storage_root/4352/"
#> 
#> [[20]]
#> [1] "https://data.scrc.uk/api/storage_root/4351/"
#> 
#> [[21]]
#> [1] "https://data.scrc.uk/api/storage_root/4350/"
#> 
#> [[22]]
#> [1] "https://data.scrc.uk/api/storage_root/4349/"
#> 
#> [[23]]
#> [1] "https://data.scrc.uk/api/storage_root/4348/"
#> 
#> [[24]]
#> [1] "https://data.scrc.uk/api/storage_root/4347/"
#> 
#> [[25]]
#> [1] "https://data.scrc.uk/api/storage_root/4346/"
#> 
#> [[26]]
#> [1] "https://data.scrc.uk/api/storage_root/4345/"
#> 
#> [[27]]
#> [1] "https://data.scrc.uk/api/storage_root/4344/"
#> 
#> [[28]]
#> [1] "https://data.scrc.uk/api/storage_root/4343/"
#> 
#> [[29]]
#> [1] "https://data.scrc.uk/api/storage_root/4342/"
#> 
#> [[30]]
#> [1] "https://data.scrc.uk/api/storage_root/4341/"
#> 
#> [[31]]
#> [1] "https://data.scrc.uk/api/storage_root/4340/"
#> 
#> [[32]]
#> [1] "https://data.scrc.uk/api/storage_root/4339/"
#> 
#> [[33]]
#> [1] "https://data.scrc.uk/api/storage_root/4338/"
#> 
#> [[34]]
#> [1] "https://data.scrc.uk/api/storage_root/4337/"
#> 
#> [[35]]
#> [1] "https://data.scrc.uk/api/storage_root/4336/"
#> 
#> [[36]]
#> [1] "https://data.scrc.uk/api/storage_root/4335/"
#> 
#> [[37]]
#> [1] "https://data.scrc.uk/api/storage_root/4334/"
#> 
#> [[38]]
#> [1] "https://data.scrc.uk/api/storage_root/4333/"
#> 
#> [[39]]
#> [1] "https://data.scrc.uk/api/storage_root/4332/"
#> 
#> [[40]]
#> [1] "https://data.scrc.uk/api/storage_root/4331/"
#> 
#> [[41]]
#> [1] "https://data.scrc.uk/api/storage_root/4330/"
#> 
#> [[42]]
#> [1] "https://data.scrc.uk/api/storage_root/4329/"
#> 
#> [[43]]
#> [1] "https://data.scrc.uk/api/storage_root/4328/"
#> 
#> [[44]]
#> [1] "https://data.scrc.uk/api/storage_root/4327/"
#> 
#> [[45]]
#> [1] "https://data.scrc.uk/api/storage_root/4326/"
#> 
#> [[46]]
#> [1] "https://data.scrc.uk/api/storage_root/4325/"
#> 
#> [[47]]
#> [1] "https://data.scrc.uk/api/storage_root/4324/"
#> 
#> [[48]]
#> [1] "https://data.scrc.uk/api/storage_root/4323/"
#> 
#> [[49]]
#> [1] "https://data.scrc.uk/api/storage_root/4322/"
#> 
#> [[50]]
#> [1] "https://data.scrc.uk/api/storage_root/4321/"
#> 
#> [[51]]
#> [1] "https://data.scrc.uk/api/storage_root/4320/"
#> 
#> [[52]]
#> [1] "https://data.scrc.uk/api/storage_root/4319/"
#> 
#> [[53]]
#> [1] "https://data.scrc.uk/api/storage_root/4318/"
#> 
#> [[54]]
#> [1] "https://data.scrc.uk/api/storage_root/4317/"
#> 
#> [[55]]
#> [1] "https://data.scrc.uk/api/storage_root/4316/"
#> 
#> [[56]]
#> [1] "https://data.scrc.uk/api/storage_root/4315/"
#> 
#> [[57]]
#> [1] "https://data.scrc.uk/api/storage_root/4314/"
#> 
#> [[58]]
#> [1] "https://data.scrc.uk/api/storage_root/4313/"
#> 
#> [[59]]
#> [1] "https://data.scrc.uk/api/storage_root/4312/"
#> 
#> [[60]]
#> [1] "https://data.scrc.uk/api/storage_root/4311/"
#> 
#> [[61]]
#> [1] "https://data.scrc.uk/api/storage_root/4310/"
#> 
#> [[62]]
#> [1] "https://data.scrc.uk/api/storage_root/4309/"
#> 
#> [[63]]
#> [1] "https://data.scrc.uk/api/storage_root/4308/"
#> 
#> [[64]]
#> [1] "https://data.scrc.uk/api/storage_root/4307/"
#> 
#> [[65]]
#> [1] "https://data.scrc.uk/api/storage_root/4306/"
#> 
#> [[66]]
#> [1] "https://data.scrc.uk/api/storage_root/4305/"
#> 
#> [[67]]
#> [1] "https://data.scrc.uk/api/storage_root/4304/"
#> 
#> [[68]]
#> [1] "https://data.scrc.uk/api/storage_root/4303/"
#> 
#> [[69]]
#> [1] "https://data.scrc.uk/api/storage_root/4302/"
#> 
#> [[70]]
#> [1] "https://data.scrc.uk/api/storage_root/4301/"
#> 
#> [[71]]
#> [1] "https://data.scrc.uk/api/storage_root/4300/"
#> 
#> [[72]]
#> [1] "https://data.scrc.uk/api/storage_root/4299/"
#> 
#> [[73]]
#> [1] "https://data.scrc.uk/api/storage_root/4298/"
#> 
#> [[74]]
#> [1] "https://data.scrc.uk/api/storage_root/4297/"
#> 
#> [[75]]
#> [1] "https://data.scrc.uk/api/storage_root/4296/"
#> 
#> [[76]]
#> [1] "https://data.scrc.uk/api/storage_root/4295/"
#> 
#> [[77]]
#> [1] "https://data.scrc.uk/api/storage_root/4294/"
#> 
#> [[78]]
#> [1] "https://data.scrc.uk/api/storage_root/4293/"
#> 
#> [[79]]
#> [1] "https://data.scrc.uk/api/storage_root/4292/"
#> 
#> [[80]]
#> [1] "https://data.scrc.uk/api/storage_root/4052/"
#> 
#> [[81]]
#> [1] "https://data.scrc.uk/api/storage_root/2968/"
#> 
#> [[82]]
#> [1] "https://data.scrc.uk/api/storage_root/2967/"
#> 
#> [[83]]
#> [1] "https://data.scrc.uk/api/storage_root/2966/"
#> 
#> [[84]]
#> [1] "https://data.scrc.uk/api/storage_root/870/"
#> 
#> [[85]]
#> [1] "https://data.scrc.uk/api/storage_root/465/"
#> 
#> [[86]]
#> [1] "https://data.scrc.uk/api/storage_root/463/"
#> 
#> [[87]]
#> [1] "https://data.scrc.uk/api/storage_root/410/"
#> 
#> [[88]]
#> [1] "https://data.scrc.uk/api/storage_root/408/"
#> 
#> [[89]]
#> [1] "https://data.scrc.uk/api/storage_root/406/"
#> 
#> [[90]]
#> [1] "https://data.scrc.uk/api/storage_root/405/"
#> 
#> [[91]]
#> [1] "https://data.scrc.uk/api/storage_root/373/"
#> 
#> [[92]]
#> [1] "https://data.scrc.uk/api/storage_root/372/"
#> 
#> [[93]]
#> [1] "https://data.scrc.uk/api/storage_root/299/"
#> 
#> [[94]]
#> [1] "https://data.scrc.uk/api/storage_root/298/"
#> 
#> [[95]]
#> [1] "https://data.scrc.uk/api/storage_root/203/"
#> 
#> [[96]]
#> [1] "https://data.scrc.uk/api/storage_root/56/"
#> 
#> [[97]]
#> [1] "https://data.scrc.uk/api/storage_root/15/"
#> 
#> [[98]]
#> [1] "https://data.scrc.uk/api/storage_root/14/"
#> 
#> [[99]]
#> [1] "https://data.scrc.uk/api/storage_root/11/"
#> 
#> [[100]]
#> [1] "https://data.scrc.uk/api/storage_root/10/"
#> 

# Retrieve the URL of a particular entry in a table
get_url(table = "storage_root", list(name = "github"))

#> [1] "https://data.scrc.uk/api/storage_root/11/"