Get URL

Retrieve URLs for particular entries or all entries in a table

get_url(table, query = list())

Arguments

table
a string specifying the name of the table

query
a list containing a valid query for the table, e.g. list(field = value)

Value

Returns a string or list of strings specifying the URL or URLs of entries in a table

Examples

# Retrieve URLs for all entries in a table
get_url(table = "storage_root")

#> [[1]]
#> [1] "https://data.scrc.uk/api/storage_root/4083/"
#> 
#> [[2]]
#> [1] "https://data.scrc.uk/api/storage_root/4082/"
#> 
#> [[3]]
#> [1] "https://data.scrc.uk/api/storage_root/4081/"
#> 
#> [[4]]
#> [1] "https://data.scrc.uk/api/storage_root/4080/"
#> 
#> [[5]]
#> [1] "https://data.scrc.uk/api/storage_root/4079/"
#> 
#> [[6]]
#> [1] "https://data.scrc.uk/api/storage_root/4078/"
#> 
#> [[7]]
#> [1] "https://data.scrc.uk/api/storage_root/4077/"
#> 
#> [[8]]
#> [1] "https://data.scrc.uk/api/storage_root/4076/"
#> 
#> [[9]]
#> [1] "https://data.scrc.uk/api/storage_root/4075/"
#> 
#> [[10]]
#> [1] "https://data.scrc.uk/api/storage_root/4074/"
#> 
#> [[11]]
#> [1] "https://data.scrc.uk/api/storage_root/4073/"
#> 
#> [[12]]
#> [1] "https://data.scrc.uk/api/storage_root/4072/"
#> 
#> [[13]]
#> [1] "https://data.scrc.uk/api/storage_root/4071/"
#> 
#> [[14]]
#> [1] "https://data.scrc.uk/api/storage_root/4070/"
#> 
#> [[15]]
#> [1] "https://data.scrc.uk/api/storage_root/4069/"
#> 
#> [[16]]
#> [1] "https://data.scrc.uk/api/storage_root/4068/"
#> 
#> [[17]]
#> [1] "https://data.scrc.uk/api/storage_root/4067/"
#> 
#> [[18]]
#> [1] "https://data.scrc.uk/api/storage_root/4066/"
#> 
#> [[19]]
#> [1] "https://data.scrc.uk/api/storage_root/4065/"
#> 
#> [[20]]
#> [1] "https://data.scrc.uk/api/storage_root/4064/"
#> 
#> [[21]]
#> [1] "https://data.scrc.uk/api/storage_root/4063/"
#> 
#> [[22]]
#> [1] "https://data.scrc.uk/api/storage_root/4062/"
#> 
#> [[23]]
#> [1] "https://data.scrc.uk/api/storage_root/4061/"
#> 
#> [[24]]
#> [1] "https://data.scrc.uk/api/storage_root/4060/"
#> 
#> [[25]]
#> [1] "https://data.scrc.uk/api/storage_root/4059/"
#> 
#> [[26]]
#> [1] "https://data.scrc.uk/api/storage_root/4058/"
#> 
#> [[27]]
#> [1] "https://data.scrc.uk/api/storage_root/4057/"
#> 
#> [[28]]
#> [1] "https://data.scrc.uk/api/storage_root/4056/"
#> 
#> [[29]]
#> [1] "https://data.scrc.uk/api/storage_root/4055/"
#> 
#> [[30]]
#> [1] "https://data.scrc.uk/api/storage_root/4054/"
#> 
#> [[31]]
#> [1] "https://data.scrc.uk/api/storage_root/4053/"
#> 
#> [[32]]
#> [1] "https://data.scrc.uk/api/storage_root/4052/"
#> 
#> [[33]]
#> [1] "https://data.scrc.uk/api/storage_root/4051/"
#> 
#> [[34]]
#> [1] "https://data.scrc.uk/api/storage_root/4050/"
#> 
#> [[35]]
#> [1] "https://data.scrc.uk/api/storage_root/4049/"
#> 
#> [[36]]
#> [1] "https://data.scrc.uk/api/storage_root/4048/"
#> 
#> [[37]]
#> [1] "https://data.scrc.uk/api/storage_root/4047/"
#> 
#> [[38]]
#> [1] "https://data.scrc.uk/api/storage_root/4046/"
#> 
#> [[39]]
#> [1] "https://data.scrc.uk/api/storage_root/4045/"
#> 
#> [[40]]
#> [1] "https://data.scrc.uk/api/storage_root/4044/"
#> 
#> [[41]]
#> [1] "https://data.scrc.uk/api/storage_root/4043/"
#> 
#> [[42]]
#> [1] "https://data.scrc.uk/api/storage_root/4042/"
#> 
#> [[43]]
#> [1] "https://data.scrc.uk/api/storage_root/4041/"
#> 
#> [[44]]
#> [1] "https://data.scrc.uk/api/storage_root/4040/"
#> 
#> [[45]]
#> [1] "https://data.scrc.uk/api/storage_root/4039/"
#> 
#> [[46]]
#> [1] "https://data.scrc.uk/api/storage_root/4038/"
#> 
#> [[47]]
#> [1] "https://data.scrc.uk/api/storage_root/4037/"
#> 
#> [[48]]
#> [1] "https://data.scrc.uk/api/storage_root/4036/"
#> 
#> [[49]]
#> [1] "https://data.scrc.uk/api/storage_root/4035/"
#> 
#> [[50]]
#> [1] "https://data.scrc.uk/api/storage_root/4034/"
#> 
#> [[51]]
#> [1] "https://data.scrc.uk/api/storage_root/4033/"
#> 
#> [[52]]
#> [1] "https://data.scrc.uk/api/storage_root/4032/"
#> 
#> [[53]]
#> [1] "https://data.scrc.uk/api/storage_root/4031/"
#> 
#> [[54]]
#> [1] "https://data.scrc.uk/api/storage_root/4030/"
#> 
#> [[55]]
#> [1] "https://data.scrc.uk/api/storage_root/4029/"
#> 
#> [[56]]
#> [1] "https://data.scrc.uk/api/storage_root/4028/"
#> 
#> [[57]]
#> [1] "https://data.scrc.uk/api/storage_root/4027/"
#> 
#> [[58]]
#> [1] "https://data.scrc.uk/api/storage_root/4026/"
#> 
#> [[59]]
#> [1] "https://data.scrc.uk/api/storage_root/4025/"
#> 
#> [[60]]
#> [1] "https://data.scrc.uk/api/storage_root/4024/"
#> 
#> [[61]]
#> [1] "https://data.scrc.uk/api/storage_root/4023/"
#> 
#> [[62]]
#> [1] "https://data.scrc.uk/api/storage_root/4022/"
#> 
#> [[63]]
#> [1] "https://data.scrc.uk/api/storage_root/4021/"
#> 
#> [[64]]
#> [1] "https://data.scrc.uk/api/storage_root/4020/"
#> 
#> [[65]]
#> [1] "https://data.scrc.uk/api/storage_root/4019/"
#> 
#> [[66]]
#> [1] "https://data.scrc.uk/api/storage_root/4018/"
#> 
#> [[67]]
#> [1] "https://data.scrc.uk/api/storage_root/4017/"
#> 
#> [[68]]
#> [1] "https://data.scrc.uk/api/storage_root/4016/"
#> 
#> [[69]]
#> [1] "https://data.scrc.uk/api/storage_root/4015/"
#> 
#> [[70]]
#> [1] "https://data.scrc.uk/api/storage_root/4014/"
#> 
#> [[71]]
#> [1] "https://data.scrc.uk/api/storage_root/4013/"
#> 
#> [[72]]
#> [1] "https://data.scrc.uk/api/storage_root/4012/"
#> 
#> [[73]]
#> [1] "https://data.scrc.uk/api/storage_root/4011/"
#> 
#> [[74]]
#> [1] "https://data.scrc.uk/api/storage_root/4010/"
#> 
#> [[75]]
#> [1] "https://data.scrc.uk/api/storage_root/4009/"
#> 
#> [[76]]
#> [1] "https://data.scrc.uk/api/storage_root/4008/"
#> 
#> [[77]]
#> [1] "https://data.scrc.uk/api/storage_root/4007/"
#> 
#> [[78]]
#> [1] "https://data.scrc.uk/api/storage_root/4006/"
#> 
#> [[79]]
#> [1] "https://data.scrc.uk/api/storage_root/4005/"
#> 
#> [[80]]
#> [1] "https://data.scrc.uk/api/storage_root/4004/"
#> 
#> [[81]]
#> [1] "https://data.scrc.uk/api/storage_root/4003/"
#> 
#> [[82]]
#> [1] "https://data.scrc.uk/api/storage_root/4002/"
#> 
#> [[83]]
#> [1] "https://data.scrc.uk/api/storage_root/4001/"
#> 
#> [[84]]
#> [1] "https://data.scrc.uk/api/storage_root/4000/"
#> 
#> [[85]]
#> [1] "https://data.scrc.uk/api/storage_root/3999/"
#> 
#> [[86]]
#> [1] "https://data.scrc.uk/api/storage_root/3998/"
#> 
#> [[87]]
#> [1] "https://data.scrc.uk/api/storage_root/3997/"
#> 
#> [[88]]
#> [1] "https://data.scrc.uk/api/storage_root/3996/"
#> 
#> [[89]]
#> [1] "https://data.scrc.uk/api/storage_root/3995/"
#> 
#> [[90]]
#> [1] "https://data.scrc.uk/api/storage_root/3994/"
#> 
#> [[91]]
#> [1] "https://data.scrc.uk/api/storage_root/3993/"
#> 
#> [[92]]
#> [1] "https://data.scrc.uk/api/storage_root/3992/"
#> 
#> [[93]]
#> [1] "https://data.scrc.uk/api/storage_root/3991/"
#> 
#> [[94]]
#> [1] "https://data.scrc.uk/api/storage_root/3990/"
#> 
#> [[95]]
#> [1] "https://data.scrc.uk/api/storage_root/3989/"
#> 
#> [[96]]
#> [1] "https://data.scrc.uk/api/storage_root/3988/"
#> 
#> [[97]]
#> [1] "https://data.scrc.uk/api/storage_root/3987/"
#> 
#> [[98]]
#> [1] "https://data.scrc.uk/api/storage_root/3986/"
#> 
#> [[99]]
#> [1] "https://data.scrc.uk/api/storage_root/3985/"
#> 
#> [[100]]
#> [1] "https://data.scrc.uk/api/storage_root/3984/"
#> 

# Retrieve the URL of a particular entry in a table
get_url(table = "storage_root", list(name = "github"))

#> [1] "https://data.scrc.uk/api/storage_root/11/"