Get URL

Retrieve URLs for particular entries or all entries in a table

get_url(table, query = list())

Arguments

table
a string specifying the name of the table

query
a list containing a valid query for the table, e.g. list(field = value)

Value

Returns a string or list of strings specifying the URL or URLs of entries in a table

Examples

# Retrieve URLs for all entries in a table
get_url(table = "storage_root")

#> [[1]]
#> [1] "https://data.scrc.uk/api/storage_root/4443/"
#> 
#> [[2]]
#> [1] "https://data.scrc.uk/api/storage_root/4442/"
#> 
#> [[3]]
#> [1] "https://data.scrc.uk/api/storage_root/4441/"
#> 
#> [[4]]
#> [1] "https://data.scrc.uk/api/storage_root/4440/"
#> 
#> [[5]]
#> [1] "https://data.scrc.uk/api/storage_root/4439/"
#> 
#> [[6]]
#> [1] "https://data.scrc.uk/api/storage_root/4438/"
#> 
#> [[7]]
#> [1] "https://data.scrc.uk/api/storage_root/4437/"
#> 
#> [[8]]
#> [1] "https://data.scrc.uk/api/storage_root/4436/"
#> 
#> [[9]]
#> [1] "https://data.scrc.uk/api/storage_root/4435/"
#> 
#> [[10]]
#> [1] "https://data.scrc.uk/api/storage_root/4434/"
#> 
#> [[11]]
#> [1] "https://data.scrc.uk/api/storage_root/4433/"
#> 
#> [[12]]
#> [1] "https://data.scrc.uk/api/storage_root/4432/"
#> 
#> [[13]]
#> [1] "https://data.scrc.uk/api/storage_root/4431/"
#> 
#> [[14]]
#> [1] "https://data.scrc.uk/api/storage_root/4430/"
#> 
#> [[15]]
#> [1] "https://data.scrc.uk/api/storage_root/4429/"
#> 
#> [[16]]
#> [1] "https://data.scrc.uk/api/storage_root/4428/"
#> 
#> [[17]]
#> [1] "https://data.scrc.uk/api/storage_root/4427/"
#> 
#> [[18]]
#> [1] "https://data.scrc.uk/api/storage_root/4426/"
#> 
#> [[19]]
#> [1] "https://data.scrc.uk/api/storage_root/4425/"
#> 
#> [[20]]
#> [1] "https://data.scrc.uk/api/storage_root/4424/"
#> 
#> [[21]]
#> [1] "https://data.scrc.uk/api/storage_root/4423/"
#> 
#> [[22]]
#> [1] "https://data.scrc.uk/api/storage_root/4422/"
#> 
#> [[23]]
#> [1] "https://data.scrc.uk/api/storage_root/4421/"
#> 
#> [[24]]
#> [1] "https://data.scrc.uk/api/storage_root/4420/"
#> 
#> [[25]]
#> [1] "https://data.scrc.uk/api/storage_root/4419/"
#> 
#> [[26]]
#> [1] "https://data.scrc.uk/api/storage_root/4418/"
#> 
#> [[27]]
#> [1] "https://data.scrc.uk/api/storage_root/4417/"
#> 
#> [[28]]
#> [1] "https://data.scrc.uk/api/storage_root/4416/"
#> 
#> [[29]]
#> [1] "https://data.scrc.uk/api/storage_root/4415/"
#> 
#> [[30]]
#> [1] "https://data.scrc.uk/api/storage_root/4414/"
#> 
#> [[31]]
#> [1] "https://data.scrc.uk/api/storage_root/4413/"
#> 
#> [[32]]
#> [1] "https://data.scrc.uk/api/storage_root/4412/"
#> 
#> [[33]]
#> [1] "https://data.scrc.uk/api/storage_root/4411/"
#> 
#> [[34]]
#> [1] "https://data.scrc.uk/api/storage_root/4410/"
#> 
#> [[35]]
#> [1] "https://data.scrc.uk/api/storage_root/4409/"
#> 
#> [[36]]
#> [1] "https://data.scrc.uk/api/storage_root/4408/"
#> 
#> [[37]]
#> [1] "https://data.scrc.uk/api/storage_root/4407/"
#> 
#> [[38]]
#> [1] "https://data.scrc.uk/api/storage_root/4406/"
#> 
#> [[39]]
#> [1] "https://data.scrc.uk/api/storage_root/4405/"
#> 
#> [[40]]
#> [1] "https://data.scrc.uk/api/storage_root/4404/"
#> 
#> [[41]]
#> [1] "https://data.scrc.uk/api/storage_root/4403/"
#> 
#> [[42]]
#> [1] "https://data.scrc.uk/api/storage_root/4402/"
#> 
#> [[43]]
#> [1] "https://data.scrc.uk/api/storage_root/4401/"
#> 
#> [[44]]
#> [1] "https://data.scrc.uk/api/storage_root/4400/"
#> 
#> [[45]]
#> [1] "https://data.scrc.uk/api/storage_root/4399/"
#> 
#> [[46]]
#> [1] "https://data.scrc.uk/api/storage_root/4398/"
#> 
#> [[47]]
#> [1] "https://data.scrc.uk/api/storage_root/4397/"
#> 
#> [[48]]
#> [1] "https://data.scrc.uk/api/storage_root/4396/"
#> 
#> [[49]]
#> [1] "https://data.scrc.uk/api/storage_root/4395/"
#> 
#> [[50]]
#> [1] "https://data.scrc.uk/api/storage_root/4394/"
#> 
#> [[51]]
#> [1] "https://data.scrc.uk/api/storage_root/4393/"
#> 
#> [[52]]
#> [1] "https://data.scrc.uk/api/storage_root/4392/"
#> 
#> [[53]]
#> [1] "https://data.scrc.uk/api/storage_root/4391/"
#> 
#> [[54]]
#> [1] "https://data.scrc.uk/api/storage_root/4390/"
#> 
#> [[55]]
#> [1] "https://data.scrc.uk/api/storage_root/4389/"
#> 
#> [[56]]
#> [1] "https://data.scrc.uk/api/storage_root/4388/"
#> 
#> [[57]]
#> [1] "https://data.scrc.uk/api/storage_root/4387/"
#> 
#> [[58]]
#> [1] "https://data.scrc.uk/api/storage_root/4386/"
#> 
#> [[59]]
#> [1] "https://data.scrc.uk/api/storage_root/4385/"
#> 
#> [[60]]
#> [1] "https://data.scrc.uk/api/storage_root/4384/"
#> 
#> [[61]]
#> [1] "https://data.scrc.uk/api/storage_root/4383/"
#> 
#> [[62]]
#> [1] "https://data.scrc.uk/api/storage_root/4382/"
#> 
#> [[63]]
#> [1] "https://data.scrc.uk/api/storage_root/4381/"
#> 
#> [[64]]
#> [1] "https://data.scrc.uk/api/storage_root/4380/"
#> 
#> [[65]]
#> [1] "https://data.scrc.uk/api/storage_root/4379/"
#> 
#> [[66]]
#> [1] "https://data.scrc.uk/api/storage_root/4378/"
#> 
#> [[67]]
#> [1] "https://data.scrc.uk/api/storage_root/4377/"
#> 
#> [[68]]
#> [1] "https://data.scrc.uk/api/storage_root/4376/"
#> 
#> [[69]]
#> [1] "https://data.scrc.uk/api/storage_root/4375/"
#> 
#> [[70]]
#> [1] "https://data.scrc.uk/api/storage_root/4374/"
#> 
#> [[71]]
#> [1] "https://data.scrc.uk/api/storage_root/4373/"
#> 
#> [[72]]
#> [1] "https://data.scrc.uk/api/storage_root/4372/"
#> 
#> [[73]]
#> [1] "https://data.scrc.uk/api/storage_root/4371/"
#> 
#> [[74]]
#> [1] "https://data.scrc.uk/api/storage_root/4052/"
#> 
#> [[75]]
#> [1] "https://data.scrc.uk/api/storage_root/2968/"
#> 
#> [[76]]
#> [1] "https://data.scrc.uk/api/storage_root/2967/"
#> 
#> [[77]]
#> [1] "https://data.scrc.uk/api/storage_root/2966/"
#> 
#> [[78]]
#> [1] "https://data.scrc.uk/api/storage_root/870/"
#> 
#> [[79]]
#> [1] "https://data.scrc.uk/api/storage_root/465/"
#> 
#> [[80]]
#> [1] "https://data.scrc.uk/api/storage_root/463/"
#> 
#> [[81]]
#> [1] "https://data.scrc.uk/api/storage_root/410/"
#> 
#> [[82]]
#> [1] "https://data.scrc.uk/api/storage_root/408/"
#> 
#> [[83]]
#> [1] "https://data.scrc.uk/api/storage_root/406/"
#> 
#> [[84]]
#> [1] "https://data.scrc.uk/api/storage_root/405/"
#> 
#> [[85]]
#> [1] "https://data.scrc.uk/api/storage_root/373/"
#> 
#> [[86]]
#> [1] "https://data.scrc.uk/api/storage_root/372/"
#> 
#> [[87]]
#> [1] "https://data.scrc.uk/api/storage_root/299/"
#> 
#> [[88]]
#> [1] "https://data.scrc.uk/api/storage_root/298/"
#> 
#> [[89]]
#> [1] "https://data.scrc.uk/api/storage_root/203/"
#> 
#> [[90]]
#> [1] "https://data.scrc.uk/api/storage_root/56/"
#> 
#> [[91]]
#> [1] "https://data.scrc.uk/api/storage_root/15/"
#> 
#> [[92]]
#> [1] "https://data.scrc.uk/api/storage_root/14/"
#> 
#> [[93]]
#> [1] "https://data.scrc.uk/api/storage_root/11/"
#> 
#> [[94]]
#> [1] "https://data.scrc.uk/api/storage_root/10/"
#> 
#> [[95]]
#> [1] "https://data.scrc.uk/api/storage_root/9/"
#> 
#> [[96]]
#> [1] "https://data.scrc.uk/api/storage_root/1/"
#> 

# Retrieve the URL of a particular entry in a table
get_url(table = "storage_root", list(name = "github"))

#> [1] "https://data.scrc.uk/api/storage_root/11/"